+++ src/feature/search/component/index.js
@@ -2,3 +2,8 @@ export { default as Search } from './Search/Search';
 export { default as SearchResults } from './SearchResults/SearchResults';
 export { default as SearchResultsErrorFallback } from './SearchResults/SearchResultsErrorFallback';
 export { default as SearchResultsSuspense } from './SearchResults/SearchResultsSuspense';
+
+// 시험후기 검색
+export { default as SearchExamReviews } from './SearchExamReviews/SearchExamReviews';
+export { default as SearchExamReviewsSuspense } from './SearchExamReviews/SearchExamReviewsSuspense';
+export { default as SearchExamReviewsErrorFallback } from './SearchExamReviews/SearchExamReviewsErrorFallback';
