+++ src/page/exam/index.js
@@ -1,0 +1,4 @@
+export { default as EditExamReviewPage } from './EditExamReviewPage/EditExamReviewPage';
+export { default as ExamReviewListPage } from './ExamReviewListPage/ExamReviewListPage';
+export { default as ExamReviewPage } from './ExamReviewPage/ExamReviewPage';
+export { default as WriteExamReviewPage } from './WriteExamReviewPage/WriteExamReviewPage';
