--- src/pages/ExamReviewEditPage/index.js
@@ -1,1 +1,0 @@
-export { default as ExamReviewEditPage } from './ExamReviewEditPage';
