--- src/pages/ExamReviewPage/index.js
@@ -1,1 +1,0 @@
-export { default as ExamReviewPage } from './ExamReviewPage';
