--- src/pages/ExamReviewWritePage/index.js
@@ -1,1 +1,0 @@
-export { default as ExamReviewWritePage } from './ExamReviewWritePage';
