--- src/pages/ExamReviewsPage/index.js
@@ -1,5 +1,0 @@
-export { default as ExamReviewsPage } from './ExamReviewsPage';
-
-export { default as SearchExamReviews } from './components/SearchExamReviews/SearchExamReviews';
-export { default as SearchExamReviewsSuspense } from './components/SearchExamReviews/SearchExamReviewsSuspense';
-export { default as SearchExamReviewsErrorFallback } from './components/SearchExamReviews/SearchExamReviewsErrorFallback';
